+++ packages/media/media-picker/examples/8-full-flow.tsx
@@ -44,7 +44,7 @@ export default class Example extends React.Component<{}, State> {
       },
     });
 
-    context.on('file-uploaded', file => {
+    context.on('file-added', file => {
       console.log('on file-uploaded', file);
     });