+++ .changeset/758ce3e1/changes.md
@@ -1,4 +1,4 @@
-- Add event emitter api to context + add first known event 'file-uploaded'
+- Add event emitter api to context + add first known event 'file-added'
 
 Integrators can now do:
 
@@ -7,7 +7,7 @@ import {ContextFactory, FileState} from '@atlaskit/media-core'
 
 const context = ContextFactory.create();
 
-context.on('file-uploaded', (file: FileState) => {
+context.on('file-added', (file: FileState) => {
   console.log(file.id)
 })
 ```